# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIForge Technical Suite — AI-powered log intelligence and observability platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DashboardSummary {
  logsIngested24h: number;
  logsPerSecond: number;
  anomaliesDetected24h: number;
  activeAlerts: number;
  criticalAlerts: number;
  servicesMonitored: number;
  servicesHealthy: number;
  servicesDegraded: number;
  servicesDown: number;
  mttdMinutes: number;
  mttdReductionPct: number;
  falsePositiveRate: number;
  falsePositiveReductionPct: number;
  engineerHoursSavedWeekly: number;
  patternsDiscovered: number;
  predictiveIssuesIdentified: number;
}

export interface LogVolumePoint {
  timestamp: string;
  debug: number;
  info: number;
  warn: number;
  error: number;
  fatal: number;
}

export interface AnomalyTrendPoint {
  timestamp: string;
  score: number;
  upperBound: number;
  lowerBound: number;
  threshold: number;
}

export interface MttdPoint {
  week: string;
  mttdMinutes: number;
  baselineMinutes: number;
}

export interface SeveritySlice {
  level: string;
  count: number;
  percentage: number;
}

export interface ServiceErrorRank {
  service: string;
  errors: number;
  warnings: number;
  deltaPct: number;
}

export interface ThroughputPoint {
  timestamp: string;
  eventsPerSecond: number;
}

export type LogEventLevel = (typeof LogEventLevel)[keyof typeof LogEventLevel];

export const LogEventLevel = {
  debug: "debug",
  info: "info",
  warn: "warn",
  error: "error",
  fatal: "fatal",
} as const;

export interface LogEvent {
  id: string;
  timestamp: string;
  level: LogEventLevel;
  service: string;
  host: string;
  traceId: string;
  message: string;
  anomalyScore: number;
}

export type AnomalySeverity =
  (typeof AnomalySeverity)[keyof typeof AnomalySeverity];

export const AnomalySeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type AnomalyStatus = (typeof AnomalyStatus)[keyof typeof AnomalyStatus];

export const AnomalyStatus = {
  open: "open",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;

export interface Anomaly {
  id: string;
  detectedAt: string;
  service: string;
  signature: string;
  severity: AnomalySeverity;
  anomalyScore: number;
  affectedRequests: number;
  status: AnomalyStatus;
  algorithm: string;
}

export interface AnomalyDetail {
  anomaly: Anomaly;
  sampleLogs: LogEvent[];
  timeline: AnomalyTrendPoint[];
}

export interface LogPattern {
  id: string;
  template: string;
  cluster: string;
  occurrences: number;
  services: string[];
  firstSeen: string;
  lastSeen: string;
  trendPct: number;
  confidence: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  active: "active",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;

export interface Alert {
  id: string;
  triggeredAt: string;
  title: string;
  service: string;
  severity: AlertSeverity;
  status: AlertStatus;
  affectedUsers: number;
  suggestedAction: string;
  confidence: number;
}

export interface RootCauseHop {
  service: string;
  component: string;
  evidence: string;
  timestamp: string;
}

export interface AlertDetail {
  alert: Alert;
  rootCause: RootCauseHop[];
  relatedLogs: LogEvent[];
}

export interface ForecastPoint {
  timestamp: string;
  /** @nullable */
  actual: number | null;
  forecast: number;
  upperBound: number;
  lowerBound: number;
}

export interface Prediction {
  id: string;
  service: string;
  component: string;
  predictedFailureAt: string;
  leadTimeMinutes: number;
  confidence: number;
  signal: string;
  recommendedAction: string;
  forecast: ForecastPoint[];
}

export type ServiceStatus = (typeof ServiceStatus)[keyof typeof ServiceStatus];

export const ServiceStatus = {
  healthy: "healthy",
  degraded: "degraded",
  down: "down",
} as const;

export interface Service {
  id: string;
  name: string;
  environment: string;
  status: ServiceStatus;
  instances: number;
  errorRate: number;
  p95LatencyMs: number;
  logsPerMinute: number;
  openAlerts: number;
}

export type GetLogVolumeTimeseriesParams = {
  window?: GetLogVolumeTimeseriesWindow;
};

export type GetLogVolumeTimeseriesWindow =
  (typeof GetLogVolumeTimeseriesWindow)[keyof typeof GetLogVolumeTimeseriesWindow];

export const GetLogVolumeTimeseriesWindow = {
  "1h": "1h",
  "6h": "6h",
  "24h": "24h",
  "7d": "7d",
} as const;

export type ListLogsParams = {
  level?: ListLogsLevel;
  service?: string;
  search?: string;
  /**
   * @minimum 1
   * @maximum 500
   */
  limit?: number;
};

export type ListLogsLevel = (typeof ListLogsLevel)[keyof typeof ListLogsLevel];

export const ListLogsLevel = {
  all: "all",
  debug: "debug",
  info: "info",
  warn: "warn",
  error: "error",
  fatal: "fatal",
} as const;