# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIForge Technical Suite — AI-powered log intelligence and observability platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /anomalies/{id}:
    get:
      operationId: getAnomaly
      tags: [anomalies]
      summary: Get a single anomaly with sample log lines
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Anomaly detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnomalyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get an alert with root cause analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AnomalyTrendPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        score:
          type: number
        upperBound:
          type: number
        lowerBound:
          type: number
        threshold:
          type: number
      required: [timestamp, score, upperBound, lowerBound, threshold]
...
    LogEvent:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warn, error, fatal]
        service:
          type: string
        host:
          type: string
        traceId:
          type: string
        message:
          type: string
        anomalyScore:
          type: number
      required: [id, timestamp, level, service, host, traceId, message, anomalyScore]
    Anomaly:
      type: object
      properties:
        id:
          type: string
        detectedAt:
          type: string
          format: date-time
        service:
          type: string
        signature:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        anomalyScore:
          type: number
        affectedRequests:
          type: integer
        status:
          type: string
          enum: [open, acknowledged, resolved]
        algorithm:
          type: string
      required:
        - id
        - detectedAt
        - service
        - signature
        - severity
        - anomalyScore
        - affectedRequests
        - status
        - algorithm
    AnomalyDetail:
      type: object
      properties:
        anomaly:
          $ref: "#/components/schemas/Anomaly"
        sampleLogs:
          type: array
          items:
            $ref: "#/components/schemas/LogEvent"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/AnomalyTrendPoint"
      required: [anomaly, sampleLogs, timeline]
...
    Alert:
      type: object
      properties:
        id:
          type: string
        triggeredAt:
          type: string
          format: date-time
        title:
          type: string
        service:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, acknowledged, resolved]
        affectedUsers:
          type: integer
        suggestedAction:
          type: string
        confidence:
          type: number
      required:
        - id
        - triggeredAt
        - title
        - service
        - severity
        - status
        - affectedUsers
        - suggestedAction
        - confidence
    RootCauseHop:
      type: object
      properties:
        service:
          type: string
        component:
          type: string
        evidence:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [service, component, evidence, timestamp]
    AlertDetail:
      type: object
      properties:
        alert:
          $ref: "#/components/schemas/Alert"
        rootCause:
          type: array
          items:
            $ref: "#/components/schemas/RootCauseHop"
        relatedLogs:
          type: array
          items:
            $ref: "#/components/schemas/LogEvent"
      required: [alert, rootCause, relatedLogs]
...