# lib/api-zod/src/generated/types/anomalyTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIForge Technical Suite — AI-powered log intelligence and observability platform
 * OpenAPI spec version: 0.1.0
 */

export interface AnomalyTrendPoint {
  timestamp: Date;
  score: number;
  upperBound: number;
  lowerBound: number;
  threshold: number;
}