# lib/api-zod/src/generated/types/rootCauseHop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AIForge Technical Suite — AI-powered log intelligence and observability platform
 * OpenAPI spec version: 0.1.0
 */

export interface RootCauseHop {
  service: string;
  component: string;
  evidence: string;
  timestamp: Date;
}